tree sprite + missing texture
Temporarily added the existing tree sprite from online to the project. It isn't the image I will use down the road, but it is important to keep track of it somewhere along the way.

Added a missing texture sprite! It will be used when the source for a sprite can't be found for one reason or another.