Redesigned Tree Sprite
Redesigned the sprite for trees! I was using a temporary texture I found online just to get trees working from the beginning, so now I am using my own texture that I made all myself! It took a few iterations to get it to where I wanted it to be, and I kept each step along the way inside of tree.psd, and I made it with Photopea.