Temp Drawing Updates + Missing T Alignments
Changes:
- Made it so the missing texture sprite is hard-coded into the game
functionality. This will prevent the game from completely crashing if no
textures are found at all. I also shrunk the missing texture sprite to
be a 2x2 image, instead of 16x16. This didn't work as expected with
other texture use-cases, like in the player's hand for items, so I am
working on rewriting the display logic for those, because they are just
magically working right now, since all items are 16x16 sprites. The
current functionality starts to break item rendering if the item sprite
is more or less than 16x16, while it should work for any image.