Offline Support!
Thanks to the reworks I made in loading the game sprites, the Service Worker should now work correctly! (No pun intended, at first, then I thought why not)

The main issue was just with Safari (Chromium worked just fine, and I can't remember if I tested Firefox or not), which didn't seem to be working with the old sprite loading and Service Worker handling, so I will have to test if these new changes do in fact make it work correct. So, this commit may dissappear for a little while if it doesn't work :P