Player Ground Shadow: Bitmap Sprite
After experimenting on other browsers, to try out the new ground shadow for players, I found that the major browsers each render SVGs in a 2D canvas context differently from each other, with no one consistent behavior. The intended behavior I was looking for was how Chrome renders it, then I found out that Safari calculates a few more extra pixels into the rasterized version when using `shape-rendering="crispEdges"`, and Firefox wasn't rendering the sprite at the correct dimentions, even with all of the necessary information applied to the shape.
Because of these outcomes, I made a bitmap version of the desired outcome, and the game will use that sprite instead of the scalable vector one.