Evorp Pizza Example
Showed the project to the team members at Classic Faithful, and they helped me with some guidance on how to make retextures for older sprites! Evorp made this redesign for the Pizza sprite. I will try to see what redesigns I can come up with to improve my textures for the game.