HUD Safe Area Insets
Fixes:
- Added a missing check for the HUD's bottom padding when on devices with safe area insets. Now the player is correctly drawn in the center of the renderer if there is a bottom safe area inset. I still need to add checks for this on the other three sides of the viewport, as I think the same issue could happen with those too.

For the next update, I want to look into clearer scaling for the renderer. Right now I am making the canvas's width and height a quarter of the space it takes up, and that removes a quarter of the possible pixels in the display space it takes up. So there could be four times more resolution available! It wouldn't get used very often, but it would be helpful for a few things, like rendering items at an angle, which is something that doesn't look the best as it is right now.