Input Detection Overlook
Fixed an overlook in the logic that auto-updates the Flatlands appearance mode to touch / not touch when inputs from each input type occur. In better words, now the D-Pad and the Debug menu button will be shown on screen after a touch event is fired, and the opposite will happen if a key event is fired. This was already a feature before, but I forgot to change the method calls to use the new appearance toggle function. It also didn't silently work "correctly" because I changed the Root .touch-device class to .touch instead, preventing the old logic from still working.

Added extra breakpoints for the hotbar scaling to apply vertically like it does horizontally. This means the HUD will fit on shorter screens/viewports better!