Module Knots
Tying the modules together in a few missing spots. Because of the finnicky way the modules were loaded before, I was relying on the implicit naming of variables which could be accessed in child modules, from the main module without importing them directly. This adds imports and exports for those variables, now making it explicit what variables come from where.

So say `app.js` had a variable called `debug_toggle`, and it's a checkbox element. `app.js` then imports the file `input.js`. I was accessing the `debug_toggle` variable reference implicitly by name over in `input.js`, since it's a child module to `app.js`, which imported `input.js`. This works afterall, but it relies on which module imported which, first. This is what caused all of the "can't be accessed before declared" errors, because it requires the modules to be run in a certain order, which shouldn't be the case.

These imports/exports for this update didn't run into any ordering issues, thankfully. I wanted to commit these first, since there are ones like `ItemSlotElement.js`, which when importing values into other modules from, it breaks the module reading order somewhere along the line.

```
Found 97 errors in 6 files.

Errors  Files
     1  src/Entity.ts:7
     1  src/Flatlands.ts:44
    14  src/ItemSlotElement.ts:18
    67  src/Player.ts:13
     9  src/Tree.ts:13
     5  src/app.ts:64
```