Property Types
Defined types for the game state classes! This removed a lot of the existing errors.

The only thing next to rework is to fix up the types for the `ItemSlotElement` class, as they are all over the place, and to fix the module loading order so I will be able to import `ItemSlotElement` into other modules, as that seems to be what's breaking the module loading order. Almost there!

I'm actually surprised how fast that was to type-out! TypeScript made errors really easy to find (that's what it does, but yeah... haha), and that makes it obvious what needs adjusting, most of the time at least. Now that everything will be typed out, it's much safer to make changes, as it will let you know if something you did breaks the existing type flow.

I'm realling looking forwards to making types and interfaces for the game features! I didn't think about that until I started this re-typing, and that's a huge benefit that comes along with this! The built-in game objects will be smart now, at least from the IDE perspective. I can safely match the JSON courterpart to the corresponding class implementation. This is so neat! Ok, I love TypeScript :)

```
Found 8 errors in 2 files.

Errors  Files
     7  src/ItemSlotElement.ts:19
     1  src/app.ts:64
```