Hotbar Slots Type
Finally got the hotbar slots type working! I had to use a type import for it to work though, haha.

I'm not exactly sure about the module loading order, except that when importing the ItemSlotElement module, and not importing anything, it loads fine.

I'm going to leave that for the main rework, as that will remove a lot of the redundancies in the codebase, including the module issue. I'm going to move a lot of things into separate components, like the HUD elements being stateful and Web Component based, a lot of nicer things like that :)

https://stackoverflow.com/questions/35551366/what-is-the-defined-execution-order-of-es6-imports

```
Found 7 errors in the same file, starting at: src/ItemSlotElement.ts:19
```