Flatlands Album Player Embed!
Release the album with the in-game music for Flatlands! Added a Bandcamp embed to try things out initially. This is still on the experimental TypeScript branch, so this won't be available on the live game for a little bit until I get this version running on there. May move down a level to JSDoc typing to get that working a bit simpler.

Lots of things since the last update (to Flatlands)! Trying to get my albums up and out into the wild, been writing more albums alongside that, and NBTify has made a ton of progress too. Not sure if it was even called NBTify since the last time I worked on this repo. Got a new Mac too! Writing this on there, and I uploaded the album on here too. Super awesome so far :O