DPad Component Module
Goofed! Tried to get the D-Pad to work on mobile for about 10 minutes or so, only to find that I simply just hadn't remembered to import it into the main app module XD

Organized the imports a bit just to help make sure I have everything imported for sure now.