Album Artwork Setup
This was a little snippet I set up to get a nice screenshot for the soundtrack's album art. I've been indecisive on what the art for this album should be, this one's been very hard to figure out how I want it to look. But I think I have it now. I found an old copy of 'Dangit Josh!' on my phone, which I had exported when I was originally making the album a while ago. This was what the screenshot looked like for the placeholder artwork I had used, and I realized I actually did like using the game's aesthetic for the artwork too. I think the problem before was just that I had made it too zoomed out, so you couldn't see the pixels as much, and that's a big part of the album, the classic retro game feel! And I'm not going to put any labels or anything on the album this time around, I think I'm going into a new era in that regard. I think I'm getting a little bit closer to learning what I'm looking for on the covers for these next albums coming up :)

Oh yeah, and the browser is zoomed in to 250% for the canvas to be the right size to take a screenshot. And the window is also in Split View (snapped, not fullscreen) with DevTools open below it (how I did all of this 'hacking' to make it work :P). I think those have to be there for you to get the right proportions when looking at this again in the future. So, to sum those up, it looks like the canvas is a square of 184 CSS pixels to be rendered at the right game size. That should help describe the layout a bit better (retro game pun unintended, initially).