Feature Typings
I'm going to work on re-integrating the feature definitions back into the game, as right now they are un-typed, and it's making it hard to tell where parts of the codebase are being defined. I will find a new, better way to implement add-ons, maybe using JS instead of plain JSON objects. I like that MC does that, but since Flatlands is written in JS anyways, might as well just provide the internal workings of the game's implementation to the 'Add-On API' so you can actually just mod the game using ESM!