Feature Typings - Stable
Meant to commit this yesterday. Have a few updates to the tsconfig that I was going to include here too, which held back not finishing this before just pushing it to production. Realized when looking at it again, those should probably come as separate changes apart from the ones of pushing the feature typings to production.