TS Service Worker + Updated Tooling!
Unrelated to Flatlands, I learned today that with Safari 17 on macOS Monterey, desktop now has support for `navigator.standalone`, which unfortunately breaks the detection I currently use to distinguish between desktop and mobile iOS devices. So I need to find a new workaround for that, as now macOS can be detected as an iPad, which isn't good.

These tooling updates to Flatlands are from my recent updates to STE and Dovetail, which I remembered I hadn't brought over here yet.