Hud Panel Tag Usage
Moving away from using custom HTML tags, in favor of eventually Solid JSX components, and just plain HTML. I still like that concept for a vanilla stack, it just doesn't play as nice for type-safe JSX declarations without a few extra boilerplate. It's not too much to have that, but it's not really necessary in reality, unless it's already for an existing Web Component or builtin HTML tag for example.