Flatlands State Class + ESM SW!
Moved away from the Flatlands state class! Now plain signals are used for state management! The Service Worker installation management is now just inlined as plain setup code as well, rather than with a method call.

Also looked into moving the Service Worker over to an ESM script, and it works great! I was worried that it's not supported cross-browser as much, but nowadays it's already been around for a year everywhere, the only previous holdup was Firefox for the longest time. Chrome and Safari supported it for a while I think.

This thankfully makes it so there's a single source of truth for where the version of the app comes from, just the `package.json`! Extra great.

The Puzzle/Snuggles (puzzle wasn't as weird as I remember, and snuggles I liked even more this time around, now being more familiar with the other calm works by Devy).

https://caniuse.com/?search=worker%20module