Proper Ref Usage!
Heck yeah!!! Been trying to debug this one the last few days. Rather than referencing the module-relative variables themselves in the refs, I use a setter function that assigns the refs to them from the App component itself.

I had to move the definitions for the refs to the top of the file as well. I think I realized that's the big problem with why the app is having so much trouble, trying to reference values before they are defined. I need to restructure the script so variables are properly defined at the top of the module. I think that's why I was having so much trouble with the ESM setup before, because I was implicitly using that to my advantage, scoping variables to the top of the module scope without intentionally doing so.