Dependencies Restructuring
Figured out the kerfuffle! So it seems to be because `definitions.js` was relying on the Canvas context to complete, which needs the Canvas itself, which is part of the render process, but the render process needs the definitions, so it's a complex cycle that ESM was figuring out for me, because it's ability to handle recursive calls. I'm not sure this is the only issue, but restructuring things seems to make this be the only issue, that I'm aware of so far. So I instead just need to make the pattern as an effect of the render process's initialization. That will be in the next commit. This one yet again doesn't change any functionality, just reorders code block sections themselves, to figure out what needs what to run correctly in order.