Slot Props Scoping
Now I see why it's very helpful to scope specific parts of state to individual components! Let alone for reproducibility, but for debugging the dependency tree as well, since you have a specific interface that declares what a component needs to function, it isn't bound by the module it's in, it's the type of the function declaration itself. So the names of values in the function are itself scoped to the function, rather than the module that contains it. This is like modules, but in a functional level. Ideally functions shouldn't need to use side-effect neighboring values, unless they are constants I'd think. The issue was that I was using side-effect function value references, but with mutable values, which is what makes it hard to find what is running where and when. This seems like a very great, and trivial thing to catch over in Rust, if I understand that right. I still haven't completely looked into crablang yet.