Localized Game State
Now I see that I think I can turn this into an instantized game loop class thing too, this should be a piece of state itself, rather than just a recursive function stack on it's own. If I implement it with a class and events, then I can interact with the loop itself, and get lifecycle hooks out of it. That's the model I want to make for a gamepad and keyboard API. Then you can very easily hook that into the state handling for something like SolidJS! Works perfectly.