App Setup Scoping
Not sure if this breaking things by removing the `await`, but I'm gonna try it for now. Ideally I want the site to cache the game assets right as soon as it is able to, we'll see though. Maybe that doesn't make sense for the first load.

I think this means everything is very close to being modularized now, once again! Now all of the dependencies for all of the components are scoped to the components themselves, rather than with singleton module imports. That actually puts it into perspective a lot, the way I was using dependencies for game features wouldn't allow it to be created dynamically, because the modules themselves weren't being used independently. Each module would still rely on the singleton state from the adjacent module, that data should instead be instantiated from a parent instance of something, hence how the App component works now. I need to adjust things to work more modularly now though, then I can start working on things like game instances!