Proper Module Usage
Now all of the imports for the new setup are used correctly! Epic! It even runs now too. I'd only expect it to work though, lol. Since it's the same code from before anyways. But that never always works out, so extra bonus points, and thanks to it for working :)

burgie + 2x animal