Debug Component Styles
Goofed up and forgot to separate these from the Hud styles.

Trying to fix a bug with something with how I style the Hud component currently, it's breaking the consistency of how `offsetY()` gets calculated, it's not symmetrical with the original version. It's 4 pixels off. It seems to have been introduced just today, with the 'Hud Component' update.