Ref Cascading
This was the issue all along! It was the change to the Ref usage, that doesn't work with the magic that Solid provides, I think it's a transpiler issue, in that it wasn't doing the same magic I was expecting it to. It was my misunderstanding of how it works I think.