Nullish Render Checks Removal
Now that I've debugged the usage of the element refs, they will always be accessible by the time the game gets to it's first render! Heck yeah! Now I can look into straightening out where things are defined, because the timing in which they are used now lines up with when they are defined :)