Stable Merge! 🍕
Alright, I think it's time! Going to move this build over to the main branch! Lots of development has been happening here, really excited to make it the main version. It's much more maintainable, and it's helping me with understanding how to use functional frontend frameworks too. Glad I managed to decipher my own original code to be able to refactor it losslessly into the new format. Now it can allow me to add new feature much more easily!

I think this is also a nice time to mark it as Version 1.0. It's been around for some time now, and while not everything is perfect (still very far from it), this is the kind of refactor I could only have dreamed of being able to complete, and I think getting as close as I have is a big enough milestone to celebrate for this project 🎉