Game Component
Realized to enable better instantiation support, I have to separate the game logic itself, from the app logic. Makes a lot of sense! So there are the global things to setup, regardless of instances of the game, and or the game server, render, split screen, what have you. The game part is the only part that needs to be dynamically duplicatable, and the App component should still only ever have just a single instance, because it's the root component of all other components.