Rotation Dedupe
Fixed the goofy split screen rotation setup, now it properly corresponds to the device's orientation when using this on a tablet. You can lay it flat on a table, and two different people can play from either side. The rotation was just doubled-up and a bit awkward. I was rotating it when I didn't really have to.